renames to scroll tracker tweaks to styles
